Project edit form variant 1.

+++ src/main/java/com/dataart/blueprintsmanager/persistence/entity/ProjectEntity.java
@@ -18,7 +18,7 @@ public class ProjectEntity {
     private String objectName;
     private String objectAddress;
     private LocalDate releaseDate;
-    private Integer volumeNumber;
+    private Long volumeNumber;
     private String volumeName;
     private String code;
     private UserEntity designer;
